fix: 즐겨찾기 추가 및 삭제

+++ backend/uniwiki/src/main/java/com/kiwi/uniwiki/domain/user/dto/response/UserResponseDTO.java
@@ -17,7 +17,7 @@ public class UserResponseDTO {
     @Getter
     @Builder
     public static class FavoriteUniversityList{
-        private Integer universityId;
+        private Short universityId;
         private String logoUrl;
         private String universityName;
     }
